pyproject.toml
[project]
name = "software-validation-toolchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "black==25.12.0",
  "google-generativeai==0.8.3",
  "instructor==1.13.0",
  "langchain==1.2.0",
  "langchain-openai==1.1.6",
  "mypy==1.19.1",
  "nltk==3.9.2",
  "pydantic==2.12.5",
  "pytest==9.0.2",
  "python-dotenv==1.2.1",
  "pyyaml==6.0.3",
  "requests==2.32.5",
  "spacy==3.8.11",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.7",
    "basedpyright>=1.28.0",
]
